requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Werkzeug==3.0.1
Flask-Cors==4.0.0
requests==2.31.0
supabase==2.10.0
python-dotenv==1.0.0
websockets>=13.0